# Conflicts:
#	app/src/main/java/com/mnassa/di/MnassaModules.kt
#	app/src/main/java/com/mnassa/screen/chats/ChatListController.kt
#	app/src/main/res/values/dimens.xml

+++ app/src/main/java/com/mnassa/activity/MainActivity.kt
@@ -106,7 +106,7 @@ open class MainActivity : AppCompatActivity(), KodeinAware, MnassaRouter by Mnas
     private val broadcastReceiver = object : BroadcastReceiver() {
         override fun onReceive(context: Context, intent: Intent) {
             val fromName = intent.getStringExtra(MnassaFirebaseMessagingService.FROM_USER)
-            val amount = intent.getStringExtra(MnassaFirebaseMessagingService.AMOUNT).toInt()
+            val amount = intent.getStringExtra(MnassaFirebaseMessagingService.AMOUNT).toLong()
             val layout = layoutInflater.inflate(R.layout.in_out_come_toast, findViewById(R.id.toastRoot))
             layout.iocView.showView(amount, fromName)
             val toast = Toast(applicationContext)
